test cases: add a typedef to the test case
STG does not model `typedef`s without a definition, but if it did, the
definition here should be filtered out.

We cannot drop the `typedef` altogether as that would leave dangling
references. Faking a dummy definition would become problematic at the
point of type deduplication, with all such `Typedef` nodes pointing to
the same thing.

PiperOrigin-RevId: 586696207
Change-Id: I966b92b6a3f637c3fe03ad3e6de4ed6e94ec6120