metrics: emit or discard metrics immediately
This is now possible because `Metrics` has the context about whether and
where to emit metrics which means we can store all metric information
locally instead of it being split between the individual kinds of metric
and a `vector` of `variant`s.

Incomplete metrics can no longer be output and there is no need for
`std::variant`, `std::monostate`, `std::visit` and friends.

The side-effect of this simplification is that metric output order is
now the order of destruction instead of construction.

Given scoped metrics:

* `A`
   * `B`
   * `C`
* `D`

The output order has changed from:

```
A
B
C
D
```

to:

```
C
B
D
A
```

Test cases have been updated.

PiperOrigin-RevId: 598620363
Change-Id: Id016022f1028807255ac89046370c6b7af1884a3