DWARF processor: always pass short type name to type definition filter check
The name is currently only used for diagnostics and for checking for
rare built-in types when the declaration file attribute is missing.

Contrary to previous expectations, the rare built-in types can be in the
`std` namespace. The check looks for `__` as a prefix to determine
whether a type is built in.

PiperOrigin-RevId: 605634489
Change-Id: I05579facc10fbb8bd01ab9a90ed6c6a0858c97b4