order: tweak order merging to prefer second sequence
The order functionality is used in the presentation of diffs of
sequences.

If there is a source code change from

```
struct s {
  int x;
};
```

to

```
struct s {
  long y;
};
```

then STG will now present this as the removal of `x` followed by the
addition of `y` (instead of the other way around). This is how `diff(1)`
presents things.

The change also affects how permutations of sequences are presented.
With the change from

```
struct s {
  int x;
  long y;
};
```

to

```
struct s {
  long y;
  int x;
};
```

STG will now mention the change of offset of `y` before that of
`x` (instead of the other way around).

These examples seem to be in opposition, but the consistent behaviour is
that STG now preserves the order of the second sequence and greedily
emits elements from the first sequence that respect this (instead of the
other way around).

Insertions, removals and renamings are more common than reorderings.
This change pragmatically favours a more familiar presentation order for
the more common cases.

PiperOrigin-RevId: 609069142
Change-Id: I0b81cc1d9b115584241976a23921a67e5ebf1a8d