rust: model `Variant` discriminant as an optional member
The Rust compiler can choose to place the discriminant at non-zero
offset in the data layout of a variant. Moreover, the discriminant can
completely be absent in cases where there is zero or only one possible
value of the discriminant. Lastly, the Rust compiler can optimise for
space by letting the discriminant overlap with the variant members data
location by using values which would be illegal for the members.

Therefore, it makes sense to model the discriminant as an optional
`Member` node for the `Variant` node.

PiperOrigin-RevId: 634746801
Change-Id: I0a20d1766570f312c4e3415229da83b7cffe04ea