ELF reader: split Linux / userspace symbol table handling
The Linux kernel module loader just checks the ksymtab of `vmlinux` and
already-loaded modules for defined symbols when linking a freshly-loaded
module.

The ksymtab is a contiguous array, ordered by symbol name, and reflected
in the ELF `.symtab` as synthetic `__ksymtab_foo` symbols, one per
symbol `foo`.

Almost all the checks in `IsPublicFunctionOrVariable` and CFI address
remapping are not needed for kernel symbols.

Extracting undefined ksymtab symbols will require extra work to get
their CRC and namespace metadata as these are stored very differently
and are not reflected in the ELF `.symtab`.

PiperOrigin-RevId: 650199778
Change-Id: I46dbce6d8d3829918fea2b7886f48c97077eb8e5