proto writer: rework custom sort functions with forward_as_tuple
Originally we thought that `operator<=>` might be useful here. However,
in C++, its usefulness is largely restricted to class-level definitions.
This may change if `std::sort` acquires overloads taking 3-way
comparisons in a future version of the language.

Instead, we use `std::forward_as_tuple` which results in readable and
efficient code.

The function `SortElfSymbolsByVersionedName` has been further simplified
firstly to let protobuf field defaults be used in case version
information is missing and secondly to not use the "is default" flag as
part of the sort key. These changes make no practical difference but
result in much simpler code.

PiperOrigin-RevId: 652880250
Change-Id: Icdfbc603128b86e09533d3ff8854e6256c9bfcae