error: split Check destructor for inlining
In the common case `Check` will succeed. Compilers are smart enough to
inline the `operator<<` and split the success / fail cases so that no
output-related work is done in the successful case.

However, the destructor can construct and `throw` an `Exception` object.
It is large enough not to be inlined. This means that the successful
case still has an unwanted function call.

Splitting the `throw` out of the destructor as `Throw` and marking this
as `[[noreturn]]` allows the destructor to be inlined. The successful
case no longer has the extra function call.

PiperOrigin-RevId: 652880856
Change-Id: Iae2d0f908519a6e82342b87f72008e7da93fd685