ELF reader: allow ELF without DWARF
Make `ElfDwarfHandle::GetDwarf` produce a warning instead of failing
with an error if DWARF information is not accessible. In this case it
returns `nullptr`, which is now accepted by `dwarf::Process`.

There may be different reasons for `dwfl_module_getdwarf` returning
`nullptr`: no debug information, split debug file or corrupted DWARF.
elfutils doesn't offer interface to tell these reasons apart, so the
best we can do for now is to warn the user.

PiperOrigin-RevId: 652882190
Change-Id: Ib60fb61ed41e2432b9f3e55bcad0f4bb777c89ec