proto_reader: ReadFromString: drop useless parameter type qualifier and add a TODO
There was a spurious `const` qualifier probably dating from when the parameter
may have been a `const std::string&`.

This `std::string_view` parameter is used to construct a `std::string` to pass
to the ProtoBuf library. This is indeed inefficient, but unfortunately
necessary, as AOSP is currently stuck on an older version of ProtoBuf. The added
`TODO` is to remind us to remove the construction once AOSP support lands.

PiperOrigin-RevId: 658793429
Change-Id: Ie330cf851ce3f197eb626eba4bb4493400705ab1