unification: clarify reference-taking in Find
In the path-splitting `Find`, references were taken to the parent node
ID and parent parent node ID. Taking references is safe because the
mapping data structure cannot be invalidated by these look-ups which
also cannot fail.

However, there was no need to take a reference to the parent parent
node ID as it is not updated in (a single iteration of) the loop body.

This code change is merely for readability, the compiled output is
identical.

PiperOrigin-RevId: 693039414
Change-Id: I826bb441de7326518d6b5e0c61cbf18ba0773fd8