comparison: replace all optional<Comparison> with Comparison{}
`Comparison` is just `pair<optional<Id>, optional<Id>>` and the
`Comparison{nullopt, nullopt}` (which is also `Comparison{}`) value
does not represent a meaningful removal, addition or change.

`optional<Comparison>{}` used to be used to represent the absence of
changes, but we now just use `Comparison{}` for that case and
eliminate a lot of unnecessary wrapping and unwrapping in `optional`.

PiperOrigin-RevId: 695342307
Change-Id: I4bd7cb0c199c7ba5a0787bf263e2b43e4eccef8d