comparison: refactor main comparison function
This change splits out the special typedef and qualifier resolution
into `CompareWithResolution`. The split allows the different
conditional branches to be turned into early returns and the function
to be flattened.

Overall, the code is more readable as there is no longer a single
`Result` object being used in multiple contexts and the compiler can
do a better job.

PiperOrigin-RevId: 695717652
Change-Id: I574042416027954df4ddcef2eb51fbf78183f5b4