DWARF processor: process `DW_TAG_type_unit`
A type unit contains the definition of a single type, plus additional
debugging information entries that may be necessary to include as part
of the definition of the type. It contains types that are shared between
compile units and referenced by `DW_AT_signature`.

Process it the same way as `DW_TAG_compile_unit` by processing all
children.

This change alone should not have any effects on tests, because types in
type units are only referenced through `DW_AT_signature`, which is
supported only in the following changes.

PiperOrigin-RevId: 717881051
Change-Id: I28a97fe47a0b2cb579b82603e743b77aa9976bc5