DWARF processor: support `DW_AT_signature` declarations
Use `DW_AT_signature` to get reference to the full type of structure and
enum declarations.

Declarations with `DW_AT_signature` don't have any other attributes, but
we need at least a name to create unifiable forward declaration. So we
get name from the full type instead.

Pass ELF reader's unification helper to DWARF processor to mark each
declaration with a signature to be unified with the full type.

PiperOrigin-RevId: 717883423
Change-Id: I57283768076fd593c27ad3e713058f814f85e5de