format: support enumerators and discriminators of arbitrary size
The external model of numerical constants is now multi-limb and can
represent integers of arbitrary size. The values are 64-bit chunks of a
2's complement number.

Examples:

| representation | value     |
|----------------|-----------|
|                |  0        |
|  0             |  0        |
|  1             |  1        |
| -2             | -2        |
|  0,  0         |  0        |
|  3,  0         |  3        |
| -4,  0         |  2^64-4   |
|  0,  1         |  2^64     |
|  5,  1         |  2^64+5   |
| -6,  1         |  2^65-6   |
|  0, -1         | -2^64     |
|  7, -1         | -(2^64-7) |
| -8, -1         | -8        |

The STG format version has not been bumped because enumerator values in
existing STG files have the same semantics (a single 2's complement
int64, absent if 0) and will be read and written identically. Absent vs
zero discriminators are represented as before (either no value or an
explicit zero value).

PiperOrigin-RevId: 797648939
Change-Id: I89e28473c6f01a4a073414956adfb8a9f93d6f63