DWARF wrappers: handle LLVM / rustc DWARF block1 numbers
LLVM emits 128-bit enumerators and discriminators as block1 format data.
The blocks are always 128 bits (16 bytes) long, even if a shorter
representation could have been used.

Unfortunately, given the use of block1, the signedness of these values
is not specified. The 2^127 values with the top bit set are ambiguous,
without taking into consideration non-local context.

The ambiguity can be resolved arbitrarily, including by treating the
numbers as all positive or all negative or picking a dividing point.

Consistent with the behaviour for a number encoded using a DWARF dataN
format, an ambiguous block1 value is decoded as an unsigned (positive)
number and the possible signed (negative) value is emitted in a warning.

PiperOrigin-RevId: 798224724
Change-Id: I6f97cc32592f226726af24ea0f3e02260302eb52